requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# OCR
Pillow>=9.0.0
opencv-python>=4.7.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Handwriting OCR (TrOCR)
# included in transformers above

# Semantic Similarity Grading
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Optional: Ollama LLM feedback (lightweight)
requests>=2.28.0

# Utilities
rich>=13.0.0
click>=8.1.0
PyYAML>=6.0
colorama>=0.4.6
tabulate>=0.9.0
